refactor(controller): move adoption code to own file
